backend/prosto_board_volume-main/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend service for wooden boards volume detection"
requires-python = ">=3.11"

dependencies = [
    "aiohttp>=3.11.12",
    "fastapi>=0.115.8",
    "numpy>=2.2.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "pydantic-settings>=2.8.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.4",
]
